\name{stringFunctions}
\alias{\%&\%}

\title{String functions}

\description{\%&\% concatenates strings.}

\usage{
string1 \%&\% string2
}

\arguments{
  \item{string1}{A character vector}
  \item{string2}{Another character vector}
}

\value{
paste(string1, string2, sep='')
}

\examples{
'this' \%&\% 'that'
}
